+++ system/src/Grav/Common/Flex/Types/Users/UserObject.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 /**
  * @package    Grav\Common\Flex
  *
- * @copyright  Copyright (c) 2015 - 2021 Trilby Media, LLC. All rights reserved.
+ * @copyright  Copyright (c) 2015 - 2022 Trilby Media, LLC. All rights reserved.
  * @license    MIT License; see LICENSE file for details.
  */
 
@@ -30,6 +30,7 @@ use Grav\Common\Flex\Types\UserGroups\UserGroupCollection;
 use Grav\Common\Flex\Types\UserGroups\UserGroupIndex;
 use Grav\Common\User\Interfaces\UserInterface;
 use Grav\Common\User\Traits\UserTrait;
+use Grav\Common\Utils;
 use Grav\Framework\File\Formatter\JsonFormatter;
 use Grav\Framework\File\Formatter\YamlFormatter;
 use Grav\Framework\Filesystem\Filesystem;
@@ -79,6 +80,8 @@ class UserObject extends FlexObject implements UserInterface, Countable
 
     /** @var Closure|null */
     static public $authorizeCallable;
+    /** @var Closure|null */
+    static public $isAuthorizedCallable;
 
     /** @var array|null */
     protected $_uploads_original;
@@ -123,7 +126,8 @@ class UserObject extends FlexObject implements UserInterface, Countable
         // Define username if it's not set.
         if (!isset($elements['username'])) {
             $storageKey = $elements['__META']['storage_key'] ?? null;
-            if (null !== $storageKey && $key === $directory->getStorage()->normalizeKey($storageKey)) {
+            $storage = $directory->getStorage();
+            if (null !== $storageKey && method_exists($storage, 'normalizeKey') && $key === $storage->normalizeKey($storageKey)) {
                 $elements['username'] = $storageKey;
             } else {
                 $elements['username'] = $key;
@@ -136,6 +140,14 @@ class UserObject extends FlexObject implements UserInterface, Countable
         }
 
         parent::__construct($elements, $key, $directory, $validate);
+    }
+
+    public function __clone()
+    {
+        $this->_access = null;
+        $this->_groups = null;
+
+        parent::__clone();
     }
 
     /**
@@ -231,13 +243,19 @@ class UserObject extends FlexObject implements UserInterface, Countable
     }
 
     /**
+     * @param UserInterface|null $user
      * @return bool
      */
-    public function isMyself(): bool
+    public function isMyself(?UserInterface $user = null): bool
     {
-        $me = $this->getActiveUser();
+        if (null === $user) {
+            $user = $this->getActiveUser();
+            if ($user && !$user->authenticated) {
+                $user = null;
+            }
+        }
 
-        return $me && $me->authenticated && $this->username === $me->username;
+        return $user && $this->username === $user->username;
     }
 
     /**
@@ -277,8 +295,8 @@ class UserObject extends FlexObject implements UserInterface, Countable
         // Check custom application access.
         $authorizeCallable = static::$authorizeCallable;
         if ($authorizeCallable instanceof Closure) {
-            $authorizeCallable->bindTo($this);
-            $authorized = $authorizeCallable($action, $scope);
+            $callable = $authorizeCallable->bindTo($this, $this);
+            $authorized = $callable($action, $scope);
             if (is_bool($authorized)) {
                 return $authorized;
             }
@@ -628,7 +646,7 @@ class UserObject extends FlexObject implements UserInterface, Countable
             $medium = MediumFactory::fromFile($path);
             if ($medium) {
                 $media->add($path, $medium);
-                $name = basename($path);
+                $name = Utils::basename($path);
                 if ($name !== $path) {
                     $media->add($name, $medium);
                 }
@@ -683,6 +701,16 @@ class UserObject extends FlexObject implements UserInterface, Countable
      */
     protected function isAuthorizedOverride(UserInterface $user, string $action, string $scope, bool $isMe = false): ?bool
     {
+        // Check custom application access.
+        $isAuthorizedCallable = static::$isAuthorizedCallable;
+        if ($isAuthorizedCallable instanceof Closure) {
+            $callable = $isAuthorizedCallable->bindTo($this, $this);
+            $authorized = $callable($user, $action, $scope, $isMe);
+            if (is_bool($authorized)) {
+                return $authorized;
+            }
+        }
+
         if ($user instanceof self && $user->getStorageKey() === $this->getStorageKey()) {
             // User cannot delete his own account, otherwise he has full access.
             return $action !== 'delete';
@@ -787,7 +815,7 @@ class UserObject extends FlexObject implements UserInterface, Countable
                 }
 
                 // Calculate path without the retina scaling factor.
-                $realpath = $filesystem->pathname($filepath) . str_replace(['@3x', '@2x'], '', basename($filepath));
+                $realpath = $filesystem->pathname($filepath) . str_replace(['@3x', '@2x'], '', Utils::basename($filepath));
 
                 $list[$filename] = [$file, $settings];
 
@@ -898,7 +926,9 @@ class UserObject extends FlexObject implements UserInterface, Countable
     protected function getGroups()
     {
         if (null === $this->_groups) {
-            $this->_groups = $this->getUserGroups()->select((array)$this->getProperty('groups'));
+            /** @var UserGroupIndex $groups */
+            $groups = $this->getUserGroups()->select((array)$this->getProperty('groups'));
+            $this->_groups = $groups;
         }
 
         return $this->_groups;
@@ -910,7 +940,7 @@ class UserObject extends FlexObject implements UserInterface, Countable
     protected function getAccess(): Access
     {
         if (null === $this->_access) {
-            $this->getProperty('access');
+            $this->_access = new Access($this->getProperty('access'));
         }
 
         return $this->_access;
@@ -925,8 +955,6 @@ class UserObject extends FlexObject implements UserInterface, Countable
         if (!$value instanceof Access) {
             $value = new Access($value);
         }
-
-        $this->_access = $value;
 
         return $value->jsonSerialize();
     }