+++ system/src/Grav/Framework/Controller/Traits/ControllerResponseTrait.php
@@ -3,7 +3,7 @@
 /**
  * @package    Grav\Framework\Controller
  *
- * @copyright  Copyright (c) 2015 - 2021 Trilby Media, LLC. All rights reserved.
+ * @copyright  Copyright (c) 2015 - 2022 Trilby Media, LLC. All rights reserved.
  * @license    MIT License; see LICENSE file for details.
  */
 
@@ -97,7 +97,7 @@ trait ControllerResponseTrait
         $headers = $headers ?? [];
         $options = $options ?? ['force_download' => true];
 
-        $file_parts = pathinfo($filename);
+        $file_parts = Utils::pathinfo($filename);
 
         if (!isset($headers['Content-Type'])) {
             $mimetype = Utils::getMimeByExtension($file_parts['extension']);
@@ -140,9 +140,9 @@ trait ControllerResponseTrait
             $code = (int)$this->getConfig()->get('system.pages.redirect_default_code', 302);
         }
 
+        $ext = Utils::pathinfo($url, PATHINFO_EXTENSION);
         $accept = $this->getAccept(['application/json', 'text/html']);
-
-        if ($accept === 'application/json') {
+        if ($ext === 'json' || $accept === 'application/json') {
             return $this->createJsonResponse(['code' => $code, 'status' => 'redirect', 'redirect' => $url]);
         }
 
@@ -217,6 +217,7 @@ trait ControllerResponseTrait
             'code' => $code,
             'status' => 'error',
             'message' => $message,
+            'redirect' => null,
             'error' => [
                 'code' => $code,
                 'message' => $message