+++ user/plugins/page-inject/page-inject.php
@@ -10,17 +10,15 @@
 namespace Grav\Plugin;
 
 use Grav\Common\Config\Config;
+use Grav\Common\Data\Data;
 use Grav\Common\Grav;
-use Grav\Common\Helpers\Excerpts;
 use Grav\Common\Page\Interfaces\PageInterface;
 use Grav\Common\Page\Pages;
 use Grav\Common\Plugin;
 use Grav\Common\Page\Page;
 use Grav\Common\Uri;
-use Grav\Framework\Psr7\Response;
-use Grav\Framework\RequestHandler\Exception\RequestException;
+use Grav\Common\Utils;
 use Grav\Plugin\Admin\Admin;
-use Psr\Http\Message\ResponseInterface;
 use RocketTheme\Toolbox\Event\Event;
 
 class PageInjectPlugin extends Plugin
@@ -53,6 +51,8 @@ class PageInjectPlugin extends Plugin
 
         $this->enable([
             'onPageContentRaw' => ['onPageContentRaw', 0],
+            'onPagesInitialized' => ['onPagesInitialized', 0],
+            'onShortcodeHandlers' => ['onShortcodeHandlers', 0],
         ]);
     }
 
@@ -70,7 +70,7 @@ class PageInjectPlugin extends Plugin
             if (!$controller->authorizeTask('pageInject', ['admin.pages.read', 'admin.super'])) {
                 http_response_code(401);
                 $json_response = [
-                    'status'  => 'error',
+                    'status' => 'error',
                     'message' => '<i class="fa fa-warning"></i> Unable to get PageInject data',
                     'details' => 'Insufficient permissions for this user.'
                 ];
@@ -88,16 +88,135 @@ class PageInjectPlugin extends Plugin
         }
     }
 
+    public function onShortcodeHandlers()
+    {
+        $this->grav['shortcode']->registerAllShortcodes(__DIR__ . '/classes/shortcodes');
+    }
+
+    /**
+     * Add content after page content was read into the system.
+     *
+     * @param Event $event An event object, when `onPageContentRaw` is fired.
+     */
+    public function onPageContentRaw(Event $event)
+    {
+        /** @var Page $page */
+        $page = $event['page'];
+
+        /** @var Config $config */
+        $config = $this->mergeConfig($page);
+
+        if ($config->get('enabled') && $config->get('active')) {
+            // Get raw content and substitute all formulas by a unique token
+            $raw = $page->getRawContent();
+
+            // build an anonymous function to pass to `parseLinks()`
+            $function = function ($matches) use (&$page, &$config) {
+
+                $search = $matches[0];
+                $type = $matches[1];
+                $page_path = $matches[2];
+
+                preg_match('#remote://(.*?)/(.*)#', $page_path, $remote_matches);
+
+                if (isset($remote_matches[1]) && $remote_matches[2]) {
+                    $remote_injections = $config->get('remote_injections', []);
+                    $remote_url = $remote_injections[$remote_matches[1]] ?? null;
+                    if ($remote_url) {
+                        $url = $remote_url . '/?action=' . $type . '&path=/' . urlencode($remote_matches[2]);
+                        return \Grav\Common\HTTP\Response::get($url);
+                    }
+                } else {
+                    $replace = $this->getInjectedPageContent($type, $page_path, $page, $config->get('processed_content'));
+                    if ($replace) {
+                        return str_replace($search, $replace, $search);
+                    }
+                }
+                return $search;
+            };
+
+            // set the parsed content back into as raw content
+            $processed_content = $this->parseInjectLinks($raw, $function);
+            $page->setRawContent($processed_content);
+        }
+    }
+
+    public function onPagesInitialized()
+    {
+        $uri = $this->grav['uri'];
+        $type = $uri->query('action');
+        $path = $uri->query('path');
+        // Handle remote calls
+        if (in_array($type, ['page-inject','content-inject']) && isset($path)) {
+           echo $this->getInjectedPageContent($type, $path);
+           exit;
+
+        }
+    }
+
+    public function registerNextGenEditorPlugin($event) {
+        $plugins = $event['plugins'];
+
+        // page-inject
+        $plugins['css'][] = 'plugin://page-inject/nextgen-editor/plugins/page-inject/page-inject.css';
+        $plugins['js'][] = 'plugin://page-inject/nextgen-editor/plugins/page-inject/page-inject.js';
+
+        $event['plugins']  = $plugins;
+        return $event;
+    }
+
+    public static function getInjectedPageContent($type, $path, $page = null, $processed_content = null): ?string
+    {
+        $pages = Grav::instance()['pages'];
+        $page = $page ?? Grav::instance()['page'];
+
+        if (is_null($processed_content)) {
+            $header = new Data((array) $page->header());
+            $processed_content = $header->get('page-inject.processed_content') ?? Grav::instance()['config']->get('plugins.page-inject.processed_content', true);
+        }
+        preg_match('/(.*)\?template=(.*)|(.*)/i', $path, $template_matches);
+
+        $path = $template_matches[1] && $template_matches[2] ? $template_matches[1] : $path;
+        $template = $template_matches[2];
+        $replace = null;
+        $page_path = Uri::convertUrl($page, $path, 'link', false, true);
+        // Cleanup any current path (`./`) references
+        $page_path = str_replace('/./', '/', $page_path);
+
+        $inject = $pages->find($page_path);
+        if ($inject instanceof PageInterface && $inject->published()) {
+            // Force HTML to avoid issues with News Feeds
+            $inject->templateFormat('html');
+            if ($type == 'page-inject') {
+                if ($template) {
+                    $inject->template($template);
+                }
+                $inject->modularTwig(true);
+                $replace = $inject->content();
+
+            } else {
+                if ($processed_content) {
+                    $replace = $inject->content();
+                } else {
+                    $replace = $inject->rawMarkdown();
+                }
+            }
+        }
+
+        return $replace;
+    }
+
     protected function getPageInjectData()
     {
         $request = $this->grav['request'];
         $data = $request->getParsedBody();
         $page_routes = $data['routes'] ?? [];
+        $json = [];
 
         /** @var Pages $pages */
         $pages = Admin::enablePages();
 
-        foreach($page_routes as $route) {
+        foreach ($page_routes as $route) {
             /** @var PageInterface */
             $page = $pages->find($route);
 
@@ -109,7 +228,7 @@ class PageInjectPlugin extends Plugin
                 ];
             } else {
                 $data = [
-                    'status'  => 'success',
+                    'status' => 'success',
                     'message' => 'Page found',
                     'data' => [
                         'title' => $page->title(),
@@ -127,81 +246,10 @@ class PageInjectPlugin extends Plugin
         return $json;
     }
 
-    /**
-     * Add content after page content was read into the system.
-     *
-     * @param  Event  $event An event object, when `onPageContentRaw` is fired.
-     */
-    public function onPageContentRaw(Event $event)
-    {
-        /** @var Page $page */
-        $page = $event['page'];
-
-        /** @var Config $config */
-        $config = $this->mergeConfig($page);
-
-
-        if ($config->get('enabled') && $config->get('active')) {
-            // Get raw content and substitute all formulas by a unique token
-            $raw = $page->getRawContent();
-
-            // build an anonymous function to pass to `parseLinks()`
-            $function = function ($matches) use (&$page, &$twig, &$config) {
-
-                $search = $matches[0];
-                $type = $matches[1];
-                $page_path = $matches[3] ?: $matches[2];
-                $template = $matches[4];
-
-                $page_path = Uri::convertUrl($page, $page_path, 'link', false, true);
-
-                $inject = $page->find($page_path);
-                if ($inject) {
-                    // Force HTML to avoid issues with News Feeds
-                    $inject->templateFormat('html');
-                    if ($type == 'page-inject') {
-                        if ($template) {
-                            $inject->template($template);
-                        }
-                        $inject->modularTwig(true);
-                        $replace = $inject->content();
-
-                    } else {
-                        if ($config->get('processed_content')) {
-                            $replace = $inject->content();
-                        } else {
-                            $replace = $inject->rawMarkdown();
-                        }
-                    }
-
-                } else {
-                    // replace with what you started with
-                    $replace = $matches[0];
-                }
-
-                // do the replacement
-                return str_replace($search, $replace, $search);
-            };
-
-            // set the parsed content back into as raw content
-            $page->setRawContent($this->parseInjectLinks($raw, $function));
-        }
-    }
-
     protected function parseInjectLinks($content, $function)
     {
-        $regex = '/\[plugin:(content-inject|page-inject)\]\(((.*)\?template=(.*)|(.*))\)/i';
+        $regex = '/\[plugin:(content-inject|page-inject)\]\((.*)\)/i';
         return preg_replace_callback($regex, $function, $content);
     }
 
-    public function registerNextGenEditorPlugin($event) {
-        $plugins = $event['plugins'];
-
-        // page-inject
-        $plugins['css'][] = 'plugin://page-inject/nextgen-editor/plugins/page-inject/page-inject.css';
-        $plugins['js'][] = 'plugin://page-inject/nextgen-editor/plugins/page-inject/page-inject.js';
-
-        $event['plugins']  = $plugins;
-        return $event;
-    }
 }